+++ Cerebrum/modules/dns/DnsOwner.py
@@ -3,10 +3,9 @@
 from Cerebrum import Utils
 from Cerebrum.DatabaseAccessor import DatabaseAccessor
 from Cerebrum.Entity import Entity, EntityName, EntitySpread
-from Cerebrum import Errors
-from Cerebrum.modules import dns
 from Cerebrum.modules.dns.DnsConstants import _DnsZoneCode
 from Cerebrum import Group
+
 
 class MXSet(DatabaseAccessor):
     """``DnsOwner.MXSet(DatabaseAccessor)`` handles the dns_mx_set and
@@ -38,7 +37,7 @@ class MXSet(DatabaseAccessor):
     def populate(self, name):
         try:
             if not self.__in_db:
-                raise RuntimeError, "populate() called multiple times."
+                raise RuntimeError("populate() called multiple times")
         except AttributeError:
             self.__in_db = False
         self.name = name
@@ -49,7 +48,7 @@ class MXSet(DatabaseAccessor):
 
     def write_db(self):
         if not self.__updated:
-            return
+            return None
         is_new = not self.__in_db
         if is_new:
             self.mx_set_id = int(self.nextval('ip_number_id_seq'))
@@ -67,7 +66,6 @@ class MXSet(DatabaseAccessor):
             SET name=:name
             WHERE mx_set_id=:mx_set_id""", binds)
         del self.__in_db
-        
         self.__in_db = True
         self.__updated = []
         return is_new
@@ -85,7 +83,7 @@ class MXSet(DatabaseAccessor):
         self.__updated = []
 
     def find_by_name(self, name):
-        mx_set_id= self.query_1("""
+        mx_set_id = self.query_1("""
         SELECT mx_set_id
         FROM [:table schema=cerebrum name=dns_mx_set]
         WHERE name=:name""", {'name': name})
@@ -122,7 +120,7 @@ class MXSet(DatabaseAccessor):
         self.execute("""
         INSERT INTO [:table schema=cerebrum name=dns_mx_set_member] (%s)
         VALUES (%s)""" % (", ".join(binds.keys()),
-                          ", ".join([":%s" % k for k in binds.keys()])),
+                          ", ".join([":%s" % k for k in binds])),
                      binds)
 
     def del_mx_set_member(self, target_id):
@@ -163,28 +161,31 @@ class MXSet(DatabaseAccessor):
         DELETE FROM [:table schema=cerebrum name=dns_mx_set]
         WHERE mx_set_id=:mx_set_id""", {
             'mx_set_id': self.mx_set_id})
-        
+
+
 class GeneralDnsRecord(object):
     """``DnsOwner.GeneralDnsRecord(object)`` is a mix-in class for
     additional TTL-enabled data like TXT records."""
 
     def __fill_coldata(self, coldata):
         binds = coldata.copy()
-        del(binds['self'])
+        del binds['self']
         binds['field_type'] = int(binds['field_type'])
-        cols = [ ("%s" % x, ":%s" % x) for x in binds.keys() ]
+        cols = [("%s" % x, ":%s" % x) for x in binds.keys()]
         return cols, binds
 
     def get_general_dns_record(self, dns_owner_id, field_type):
-        return self.query_1("""
+        stmt = """
         SELECT ttl, data
         FROM [:table schema=cerebrum name=dns_general_dns_record]
-        WHERE dns_owner_id=:dns_owner_id AND field_type=:field_type""", locals())
+        WHERE dns_owner_id=:dns_owner_id AND field_type=:field_type"""
+        return self.query_1(stmt, locals())
 
     def add_general_dns_record(self, dns_owner_id, field_type, ttl, data):
         cols, binds = self.__fill_coldata(locals())
         self.execute("""
-        INSERT INTO [:table schema=cerebrum name=dns_general_dns_record] (%(tcols)s)
+        INSERT INTO [:table schema=cerebrum name=dns_general_dns_record]
+        (%(tcols)s)
         VALUES (%(binds)s)""" % {'tcols': ", ".join([x[0] for x in cols]),
                                  'binds': ", ".join([x[1] for x in cols])},
                      binds)
@@ -193,26 +194,58 @@ class GeneralDnsRecord(object):
                                                  'data': data})
 
     def delete_general_dns_record(self, dns_owner_id, field_type):
-        where = ['dns_owner_id=:dns_owner_id', 'field_type=:field_type']
-        where = " AND ".join(where)
-        return self.execute("""
-        DELETE FROM [:table schema=cerebrum name=dns_general_dns_record]
-        WHERE  %s """ % where, locals())
-        self._db.log_change(dns_owner_id, self.clconst.general_dns_record_del, None,
+        exists_stmt = """
+          SELECT EXISTS (
+            SELECT 1
+            FROM [:table schema=cerebrum name=dns_general_dns_record]
+            WHERE  dns_owner_id=:dns_owner_id AND
+                   field_type=:field_type
+           )
+        """
+        if not self.query_1(exists_stmt, locals()):
+            # False positive
+            return
+        delete_stmt = """
+          DELETE FROM [:table schema=cerebrum name=dns_general_dns_record]
+          WHERE dns_owner_id=:dns_owner_id AND
+                field_type=:field_type
+        """
+        self.execute(delete_stmt, locals())
+        self._db.log_change(dns_owner_id,
+                            self.clconst.general_dns_record_del,
+                            None,
                             change_params={'field_type': int(field_type)})
 
     def update_general_dns_record(self, dns_owner_id, field_type, ttl, data):
         cols, binds = self.__fill_coldata(locals())
-        self.execute("""
+        defs = {'set_defs': ", ".join(["%s=%s" % x for x in cols])}
+        exists_stmt = """
+        SELECT EXISTS (
+        SELECT 1
+        FROM [:table schema=cerebrum name=dns_general_dns_record]
+        WHERE dns_owner_id=:dns_owner_id AND
+              field_type=:field_type AND
+              data=:data AND
+             (ttl is NULL AND :ttl is NULL OR ttl=:ttl)
+        )
+        """ % defs
+        if self.query_1(exists_stmt, binds):
+            # False positive
+            return
+        update_stmt = """
         UPDATE [:table schema=cerebrum name=dns_general_dns_record]
-        SET %(defs)s
-        WHERE dns_owner_id=:dns_owner_id AND field_type=:field_type""" % {'defs': ", ".join(
-            ["%s=%s" % x for x in cols])}, binds)
-        self._db.log_change(dns_owner_id, self.clconst.general_dns_record_update,
-                            None, change_params={'field_type': int(field_type),
-                                                 'data': data})
+        SET %(set_defs)s
+        WHERE dns_owner_id=:dns_owner_id AND field_type=:field_type
+        """ % defs
+        self.execute(update_stmt, binds)
+        self._db.log_change(dns_owner_id,
+                            self.clconst.general_dns_record_update,
+                            None,
+                            change_params={'field_type': int(field_type),
+                                           'data': data})
 
-    def list_general_dns_records(self, field_type=None, dns_owner_id=None, zone=None):
+    def list_general_dns_records(self, field_type=None,
+                                 dns_owner_id=None, zone=None):
         where = ['d.dns_owner_id=gdns.dns_owner_id']
         if field_type is not None:
             field_type = int(field_type)
@@ -229,7 +262,10 @@ class GeneralDnsRecord(object):
              [:table schema=cerebrum name=dns_owner] d
         WHERE %s""" % where, locals())
 
+
 Entity_class = Utils.Factory.get("Entity")
+
+
 class DnsOwner(GeneralDnsRecord, EntityName, EntitySpread, Entity_class):
     """``DnsOwner(GeneralDnsRecord, Entity)`` primarily updates the
     DnsOwner table using the standard Cerebrum populate framework.
@@ -260,7 +296,7 @@ class DnsOwner(GeneralDnsRecord, EntityName, EntitySpread, Entity_class):
             Entity.populate(self, self.const.entity_dns_owner)
         try:
             if not self.__in_db:
-                raise RuntimeError, "populate() called multiple times."
+                raise RuntimeError("populate() called multiple times")
         except AttributeError:
             self.__in_db = False
         self.name = name
@@ -271,47 +307,64 @@ class DnsOwner(GeneralDnsRecord, EntityName, EntitySpread, Entity_class):
         assert isinstance(other, DnsOwner)
         return (self.name == other.name and
                 self.mx_set_id == other.mx_set_id)
-    
+
     def write_db(self):
         self.__super.write_db()
         if not self.__updated:
-            return
+            return None
         is_new = not self.__in_db
-
         binds = {
             'e_id': self.entity_id,
             'e_type': int(self.const.entity_dns_owner),
             'name': self.name,
             'mx_set_id': self.mx_set_id}
-
         # Do some primitive checks to assert that name is a FQDN, and
         # that the zone matches the DN.
         if not self.name[-1] == '.':
             raise ValueError("hostname must be fully qualified")
         if isinstance(self.zone, _DnsZoneCode):
-            if (self.zone.postfix is not None and
-                not self.name.endswith(self.zone.postfix)):
+            if (
+                    self.zone.postfix is not None and
+                    not self.name.endswith(self.zone.postfix)):
                 raise ValueError("Zone mismatch for %s" % self.name)
             binds['zone_id'] = self.zone.zone_id
         else:
             binds['zone_id'] = self.zone
         if is_new:
-            self.execute("""
+            insert_stmt = """
             INSERT INTO [:table schema=cerebrum name=dns_owner]
               (dns_owner_id, entity_type, mx_set_id, zone_id)
-            VALUES (:e_id, :e_type, :mx_set_id, :zone_id)""", binds)
+            VALUES (:e_id, :e_type, :mx_set_id, :zone_id)"""
+            self.execute(insert_stmt, binds)
             self.add_entity_name(self.const.dns_owner_namespace, self.name)
-            self._db.log_change(self.entity_id, self.clconst.dns_owner_add, None)
+            self._db.log_change(self.entity_id,
+                                self.clconst.dns_owner_add,
+                                None)
         else:
-            self.execute("""
-            UPDATE [:table schema=cerebrum name=dns_owner]
-            SET mx_set_id=:mx_set_id, zone_id=:zone_id
-            WHERE dns_owner_id=:e_id""", binds)
-            if 'name' in self.__updated:
-                self.update_entity_name(self.const.dns_owner_namespace, self.name)
-            self._db.log_change(self.entity_id, self.clconst.dns_owner_update, None)
+            exists_stmt = """
+              SELECT EXISTS (
+                SELECT 1
+                FROM [:table schema=cerebrum name=dns_owner]
+                WHERE dns_owner_id=:e_id AND
+                     (mx_set_id is NULL AND :mx_set_id is NULL OR
+                        mx_set_id=:mx_set_id) AND
+                      zone_id=:zone_id
+              )
+            """
+            if not self.query_1(exists_stmt, binds):
+                # True positive
+                update_stmt = """
+                  UPDATE [:table schema=cerebrum name=dns_owner]
+                  SET mx_set_id=:mx_set_id, zone_id=:zone_id
+                  WHERE dns_owner_id=:e_id"""
+                self.execute(update_stmt, binds)
+                if 'name' in self.__updated:
+                    self.update_entity_name(self.const.dns_owner_namespace,
+                                            self.name)
+                self._db.log_change(self.entity_id,
+                                    self.clconst.dns_owner_update,
+                                    None)
         del self.__in_db
-        
         self.__in_db = True
         self.__updated = []
         return is_new
@@ -342,16 +395,30 @@ class DnsOwner(GeneralDnsRecord, EntityName, EntitySpread, Entity_class):
             g.clear()
             g.find(row['group_id'])
             g.remove_member(self.entity_id)
-        self.execute("""
-        DELETE FROM [:table schema=cerebrum name=dns_owner]
-        WHERE dns_owner_id=:dns_owner_id""", {'dns_owner_id': self.entity_id})
-        self._db.log_change(self.entity_id, self.clconst.dns_owner_del, None)
+        binds = {'dns_owner_id': self.entity_id}
+        exists_stmt = """
+          SELECT EXISTS (
+            SELECT 1
+            FROM [:table schema=cerebrum name=dns_owner]
+            WHERE dns_owner_id=:dns_owner_id
+          )
+        """
+        if self.query_1(exists_stmt, binds):
+            # True positive
+            delete_stmt = """
+              DELETE FROM [:table schema=cerebrum name=dns_owner]
+              WHERE dns_owner_id=:dns_owner_id"""
+            self.execute(delete_stmt, binds)
+            self._db.log_change(self.entity_id,
+                                self.clconst.dns_owner_del,
+                                None)
         self.__super.delete()
 
     def list(self, zone=None):
         return self.search(zone=zone)
 
     def search(self, name_like=None, zone=None, fetchall=False):
+        """Search for owner"""
         where = ['d.dns_owner_id=en.entity_id']
         if name_like is not None:
             expr, name_like = self._db.sql_pattern("en.entity_name", name_like,
@@ -372,13 +439,13 @@ class DnsOwner(GeneralDnsRecord, EntityName, EntitySpread, Entity_class):
 
     def __fill_coldata(self, coldata):
         binds = coldata.copy()
-        del(binds['self'])            
-        cols = [ ("%s" % x, ":%s" % x) for x in binds.keys() ]
+        del binds['self']
+        cols = [("%s" % x, ":%s" % x) for x in binds.keys()]
         return cols, binds
-
 
     def add_srv_record(self, service_owner_id, pri, weight, port, ttl,
                        target_owner_id):
+        """add crv record"""
         cols, binds = self.__fill_coldata(locals())
         ret_value = self.execute("""
         INSERT INTO [:table schema=cerebrum name=dns_srv_record] (%(tcols)s)
@@ -388,22 +455,33 @@ class DnsOwner(GeneralDnsRecord, EntityName, EntitySpread, Entity_class):
         self._db.log_change(service_owner_id, self.clconst.srv_record_add,
                             target_owner_id)
         return ret_value
-    
 
     def delete_srv_record(self, service_owner_id, pri, weight, port,
                           target_owner_id):
+        """delete all srv records"""
         cols, binds = self.__fill_coldata(locals())
-        ret_value = self.execute("""
+        where = " AND ".join(["%s=%s" % (x[0], x[1]) for x in cols])
+        exists_stmt = """
+        SELECT EXISTS (
+        SELECT 1
+        FROM [:table schema=cerebrum name=dns_srv_record]
+        WHERE %s
+        )
+        """ % where
+        if not self.query_1(exists_stmt, binds):
+            # False positive
+            return
+        delete_stmt = """
         DELETE FROM [:table schema=cerebrum name=dns_srv_record]
-        WHERE %s""" % " AND ".join(["%s=%s" % (x[0], x[1])
-                                    for x in cols]), binds)
-        self._db.log_change(service_owner_id, self.clconst.srv_record_del,
+        WHERE %s""" % where
+        self.execute(delete_stmt, binds)
+        self._db.log_change(service_owner_id,
+                            self.clconst.srv_record_del,
                             target_owner_id)
-        return ret_value
-        
 
     def list_srv_records(self, owner_id=None, target_owner_id=None, zone=None,
                          pri=None, weight=None, port=None):
+        """List all srv records"""
         where = ['srv.target_owner_id=d_tgt.dns_owner_id',
                  'srv.service_owner_id=d_own.dns_owner_id',
                  'srv.target_owner_id=en_tgt.entity_id',
@@ -441,11 +519,11 @@ class DnsOwner(GeneralDnsRecord, EntityName, EntitySpread, Entity_class):
             'zone': zone,
             'pri': pri,
             'weight': weight,
-            'port': port} )
-
+            'port': port})
 
     # We don't support a general update_srv_record as the PK is too wide.
     def update_srv_record_ttl(self, owner_id, ttl):
+        """Update srv record (ttl only!)"""
         return self.execute("""
         UPDATE [:table schema=cerebrum name=dns_srv_record]
         SET ttl=:ttl
